Add fast path to `ArenaAllocator`
* Assembly for `ArenaAllocator.Allocate(ulong)`:

  .L0:
    mov rax, [rcx+0x18]
    lea r8, [rax+rdx]
    cmp r8, [rcx+0x10]
    ja short .L2
  .L1:
    mov rdx, [rcx+8]
    add rax, [rdx+8]
    mov [rcx+0x18], r8
    ret
  .L2:
    jmp ArenaAllocator.AllocateSlow(UInt64)

  A few variable/field had to be changed to ulong so that RyuJIT avoids
  emitting zero-extends.

* Implement a new heuristic to free pooled pages.

  If an arena is used often, it is more likely that its pages will be
  needed, so the pages are kept for longer (e.g: during PPTC rebuild or
  burst sof compilations). If is not used often, then it is more likely
  that its pages will not be needed (e.g: after PPTC rebuild or bursts
  of compilations).

+++ ARMeilleure/CodeGen/Optimizations/Optimizer.cs
@@ -105,7 +105,7 @@ namespace ARMeilleure.CodeGen.Optimizations
 
             if (buffer.Length < uses.Length)
             {
-                buffer = Allocators.Default.AllocateSpan<Operation>(uses.Length);
+                buffer = Allocators.Default.AllocateSpan<Operation>((uint)uses.Length);
             }
 
             uses.CopyTo(buffer);
