+++ ARMeilleure/IntermediateRepresentation/Operation.cs
@@ -251,7 +251,7 @@ namespace ARMeilleure.IntermediateRepresentation
             // We only need to allocate a new buffer if we're increasing the size.
             else if (newCapacity > capacity)
             {
-                list = Allocators.References.Allocate<Operand>(newCapacity);
+                list = Allocators.References.Allocate<Operand>((uint)newCapacity);
             }
 
             capacity = (ushort)newCapacity;
